first iteration of xml downloader
We need to figure out if we want the directory where the xmls are stored
to be there permanently or be deleted after they are run through the
parser.  I also need to figure out a way to deduce which builds are new.
This could be based either on time, or a value we have stored somewhere.